Split keybindingmanager from interactionhandler interactionhandler
