+++ include/openspace/engine/openspaceengine.h
@@ -62,7 +62,10 @@ class SyncEngine;
 class TimeManager;
 class WindowWrapper;
 
-namespace interaction { class InteractionHandler; }
+namespace interaction {
+    class InteractionHandler;
+    class KeyBindingManager;
+}
 namespace gui { class GUI; }
 namespace properties { class PropertyOwner; }
 namespace scripting {
@@ -124,6 +127,7 @@ public:
     WindowWrapper& windowWrapper();
     ghoul::fontrendering::FontManager& fontManager();
     interaction::InteractionHandler& interactionHandler();
+    interaction::KeyBindingManager& keyBindingManager();
     properties::PropertyOwner& globalPropertyOwner();
     scripting::ScriptEngine& scriptEngine();
     scripting::ScriptScheduler& scriptScheduler();
@@ -196,6 +200,7 @@ private:
     std::unique_ptr<ghoul::cmdparser::CommandlineParser> _commandlineParser;
     std::unique_ptr<ghoul::fontrendering::FontManager> _fontManager;
     std::unique_ptr<interaction::InteractionHandler> _interactionHandler;
+    std::unique_ptr<interaction::KeyBindingManager> _keyBindingManager;
     std::unique_ptr<scripting::ScriptEngine> _scriptEngine;
     std::unique_ptr<scripting::ScriptScheduler> _scriptScheduler;
     std::unique_ptr<VirtualPropertyManager> _virtualPropertyManager;
