+++ src/interaction/keybindingmanager_lua.inl
@@ -26,114 +26,115 @@ namespace openspace {
 
 namespace luascriptfunctions {
 /**
- * \ingroup LuaScripts
- * setOrigin():
- * Set the origin of the camera
- */
-int setOrigin(lua_State* L) {
+* \ingroup LuaScripts
+* bindKey():
+* Binds a key to Lua command to both execute locally
+* and broadcast to all clients if this node is hosting
+* a parallel connection.
+*/
+int bindKey(lua_State* L) {
     using ghoul::lua::luaTypeToString;
 
     int nArguments = lua_gettop(L);
-    if (nArguments != 1)
-        return luaL_error(L, "Expected %i arguments, got %i", 1, nArguments);
+    if (nArguments != 2 && nArguments != 3) {
+        return luaL_error(L, "Expected %i or %i arguments, got %i", 2, 3, nArguments);
+    }
 
-    const int type = lua_type(L, -1);
-    if (type != LUA_TSTRING)
-        return luaL_error(L, "Expected string, got %i", type);
+    int KeyLocation = nArguments == 3 ? -3 : -2;
+    int CommandLocation = nArguments == 3 ? -2 : -1;
+    int DocumentationLocation = -1;
 
-    std::string s = luaL_checkstring(L, -1);
+    std::string key = luaL_checkstring(L, KeyLocation);
+    std::string command = luaL_checkstring(L, CommandLocation);
 
-    SceneGraphNode* node = sceneGraphNode(s);
-    if (!node) {
-        LWARNINGC(
-            "lua.setOrigin",
-            "Could not find a node in scenegraph called '" << s << "'"
-        );
+    if (command.empty()) {
+        return luaL_error(L, "Command string is empty");
+    }
+
+    openspace::KeyWithModifier iKey = openspace::stringToKey(key);
+
+    if (iKey.key == openspace::Key::Unknown) {
+        LERRORC("lua.bindKey", "Could not find key '"<< key << "'");
         return 0;
     }
 
-    OsEng.interactionHandler().setFocusNode(node);
-
-    return 0;
-}
-
-int goToChunk(lua_State* L) {
-    using ghoul::lua::luaTypeToString;
-    
-    int nArguments = lua_gettop(L);
-    if (nArguments != 3)
-    return luaL_error(L, "Expected %i arguments, got %i", 3, nArguments);
-    
-    int x = static_cast<int>(lua_tonumber(L, 1));
-    int y = static_cast<int>(lua_tonumber(L, 2));
-    int level = static_cast<int>(lua_tonumber(L, 3));
-
-    OsEng.interactionHandler().goToChunk(x, y, level);
-
-    return 0;
-}
-
-int goToGeo(lua_State* L) {
-    using ghoul::lua::luaTypeToString;
-    
-    int nArguments = lua_gettop(L);
-    if (nArguments != 2)
-    return luaL_error(L, "Expected %i arguments, got %i", 2, nArguments);
-    
-    double latitude = static_cast<int>(lua_tonumber(L, 1));
-    double longitude = static_cast<int>(lua_tonumber(L, 2));
-    
-    OsEng.interactionHandler().goToGeo(latitude, longitude);
-    
-    return 0;
-}
-
-int restoreCameraStateFromFile(lua_State* L) {
-    using ghoul::lua::luaTypeToString;
-
-    int nArguments = lua_gettop(L);
-    if (nArguments != 1)
-        return luaL_error(L, "Expected %i arguments, got %i", 1, nArguments);
-
-    std::string cameraStateFilePath = luaL_checkstring(L, -1);
-
-    if (cameraStateFilePath.empty()) {
-        return luaL_error(L, "filepath string is empty");
+    std::string documentation;
+    if (nArguments == 3) {
+        documentation = luaL_checkstring(L, DocumentationLocation);
     }
 
-    OsEng.interactionHandler().restoreCameraStateFromFile(cameraStateFilePath);
+    OsEng.keyBindingManager().bindKey(
+        iKey.key,
+        iKey.modifier,
+        std::move(command),
+        std::move(documentation)
+    );
+        
     return 0;
 }
 
-int saveCameraStateToFile(lua_State* L) {
+/**
+* \ingroup LuaScripts
+* bindKey():
+* Binds a key to Lua command to execute only locally
+*/
+int bindKeyLocal(lua_State* L) {
     using ghoul::lua::luaTypeToString;
 
     int nArguments = lua_gettop(L);
-    if (nArguments != 1)
-        return luaL_error(L, "Expected %i arguments, got %i", 1, nArguments);
-
-    std::string cameraStateFilePath = luaL_checkstring(L, -1);
-
-    if (cameraStateFilePath.empty()) {
-        return luaL_error(L, "filepath string is empty");
+    if (nArguments != 2 && nArguments != 3) {
+        return luaL_error(L, "Expected %i or %i arguments, got %i", 2, 3, nArguments);
     }
 
-    OsEng.interactionHandler().saveCameraStateToFile(cameraStateFilePath);
+    int KeyLocation = nArguments == 3 ? -3 : -2;
+    int CommandLocation = nArguments == 3 ? -2 : -1;
+    int DocumentationLocation = -1;
+
+    std::string key = luaL_checkstring(L, KeyLocation);
+    std::string command = luaL_checkstring(L, CommandLocation);
+
+
+    if (command.empty())
+        return luaL_error(L, "Command string is empty");
+
+    openspace::KeyWithModifier iKey = openspace::stringToKey(key);
+
+    if (iKey.key == openspace::Key::Unknown) {
+        LERRORC("lua.bindKey", "Could not find key '" << key << "'");
+        return 0;
+    }
+
+    std::string documentation;
+    if (nArguments == 3) {
+        documentation = luaL_checkstring(L, DocumentationLocation);
+    }
+
+    OsEng.keyBindingManager().bindKeyLocal(
+        iKey.key,
+        iKey.modifier,
+        std::move(command),
+        std::move(documentation)
+        );
+
     return 0;
 }
 
-int resetCameraDirection(lua_State* L) {
+/**
+* \ingroup LuaScripts
+* clearKeys():
+* Clears all key bindings
+*/
+int clearKeys(lua_State* L) {
     using ghoul::lua::luaTypeToString;
 
     int nArguments = lua_gettop(L);
-    if (nArguments != 0) {
+    if (nArguments != 0)
         return luaL_error(L, "Expected %i arguments, got %i", 0, nArguments);
-    }
 
-    OsEng.interactionHandler().resetCameraDirection();
+    OsEng.keyBindingManager().resetKeyBindings();
+
     return 0;
 }
-
 
 } // namespace luascriptfunctions
